move dummy data into server/index and pass into article as page data. add test in article page tests
