+++ src/app/models/types/optimo.ts
@@ -5,6 +5,7 @@ import { TopStoryItem } from '#app/pages/ArticlePage/PagePromoSections/TopStorie
 import { LatestMedia } from '#app/pages/MediaArticlePage/PagePromoSections/LatestMediaSection/types';
 import { PageTypes } from './global';
 import { MetadataFormats, MetadataTaggings, MetadataTopics } from './metadata';
+import { Curation } from './curationData';
 
 export type OptimoBlock = {
   type: string;
@@ -130,6 +131,8 @@ export type ArticlePromo = {
 };
 
 export type SecondaryColumn = {
+  billboardCuration?: Curation;
+  multimediaCuration?: Curation;
   topStories: TopStoryItem[];
   features: object[];
   latestMedia?: LatestMedia[];
